+++ packages/core/src/oot/play/play.c
@@ -488,6 +488,12 @@ static void Play_AfterInit(PlayState* play)
     /* Signs */
     ComboPlay_SpawnExtraSigns(play);
 
+    /* Adult kakariko gate */
+    if (Config_Flag(CFG_OOT_ADULT_KAKARIKO_GATE) && play->sceneId == SCE_OOT_KAKARIKO_VILLAGE && gSave.age == AGE_ADULT)
+    {
+        Actor_Spawn(&play->actorCtx, play, ACTOR_BG_GATE_SHUTTER, 91, 400, -1350, 0, 0xf99a, 0, 0xffff);
+    }
+
     switch (play->sceneId)
     {
     case SCE_OOT_DEKU_TREE: